fix: use static cast for double => float

+++ Core/Source/Platform/Windows/WindowsPlatformUtils.cpp
@@ -6,6 +6,6 @@
 namespace Sophon {
     float Time::GetTime()
     {
-        return glfwGetTime();
+        return static_cast<float>(glfwGetTime());
     }
 }
